Suppress employee information duplication
Some employees have two positions, phones, fax numbers, etc. listed yet they are the same. In this event, rather than showing the same information twice in the employee's detail view controller, set one of the two attributes == "" so that the code will know not to show it. It will be as if it didn't exist.